Apply Mike Barker Scatter Function
By default, StarFISH uses a rather complex algorithm to add photometric
errors to the synthetic CMDs. It involves binning the AST results into
2-dimensional histograms, which are re-expressed as cumulative
functions, and these are interpolated in the CMD to the position of each
synthetic star.

As an alternative, Mike Barker came up with a more straightforward
scheme, in which the algorithm simply identifies the artificial star
which is closest to the model star in the CMD plane, and adopts the
scattering and dropout state of that artificial star. I find that this
algorithm is not only simpler, but runs faster and produces better
synthetic CMDs. Thanks for contributing it, Mike!

After installing the patchset, simply set the err_method parameter in
the synth.dat file to "2" to use it. This is the parameter whose comment
reads 0 = use AST errs; 1 = use analytic err model.