Apply Faster Error Determination
Confidence intervals on each SFH are determined by deviating from the
best-fit position in parameter space. This involves choosing a
parameter-space direction and moving in that direction until a
threshhold increase in chi-squared is reached. This is done for many
thousands of parameter-space directions. Sometimes, a single step in a
given direction changes chi-squared very slowly, so a large number of
steps is required to cross the chi-squared threshhold. This patch will
increase the step-size when it detects that the step has resulted in a
very small increase in chi-squared. This significantly reduces the time
required to determine the confidence intervals.