Apply Compact Screen Output
As StarFISH converges toward the best fit, it reports the curent
chi-squared value to the screen. This usually results in a large number
of lines being dumped to the terminal. This patch adds a C routine to
output the chi-squared values to the screen without adding newlines, so
the total output to the terminal is a much smaller number of lines.